pyproject.toml
[project]
name = "super-agent-party"
version = "v0.2.5"
description = "超级智能体派对v0.2.5 | 轻松链接一切!"
license = "AGPL-3.0"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.11.18",
    "aiosqlite>=0.21.0",
    "appdirs>=1.4.4",
    "bs4>=0.0.2",
    "docstring-parser>=0.16",
    "faiss-cpu>=1.10.0",
    "fastapi-mcp>=0.3.4",
    "fastapi>=0.115.12",
    "langchain-community>=0.3.22",
    "langchain-ollama>=0.3.2",
    "langchain-openai>=0.3.14",
    "mcp>=1.6.0",
    "nest-asyncio>=1.6.0",
    "odfpy>=1.4.1",
    "openai>=1.76.0",
    "openpyxl>=3.1.5",
    "pyinstaller>=6.13.0",
    "pypdf2>=3.0.1",
    "python-a2a[all]>=0.5.6",
    "python-docx>=1.1.2",
    "python-multipart>=0.0.20",
    "python-pptx>=1.0.2",
    "pywin32>=310 ; sys_platform == 'win32'",
    "rank-bm25>=0.2.2",
    "requests>=2.32.3",
    "shortuuid>=1.0.13",
    "striprtf>=0.0.29",
    "tavily-python>=0.7.0",
    "tzlocal>=5.3.1",
    "uvicorn[standard]>=0.34.2",
    "mem0ai>=0.1.101",
    "e2b-code-interpreter>=1.5.0",
    "qq-botpy>=1.2.1",
    "duckduckgo-search>=8.0.3",
    "langchain-google-community>=2.0.7",
    "langchain-exa>=0.3.0",
    "ollama>=0.5.1",
    "scipy>=1.16.0",
    "edge-tts>=7.0.2",
    "brotli>=1.1.0",
    "wxauto>=39.1.12; sys_platform == 'win32'",
    "wikipedia-api>=0.8.1",
    "arxiv>=2.2.0",
    "anyio>=4.9.0",
    "httpcore>=1.0.9",
    "httpx>=0.28.1",
    "ddgs>=9.5.4",
    "dify-client-python>=1.0.1",
]
